Break dependency of core.h on pipeline.h
Move the step definitions to plain.h
