+++ src/client/render/plain.cpp
@@ -19,11 +19,78 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 */
 
 #include "plain.h"
-#include "settings.h"
+#include "client/camera.h"
+#include "client/client.h"
+#include "client/clientmap.h"
+#include "client/hud.h"
+#include "client/minimap.h"
+#include "client/shadows/dynamicshadowsrender.h"
 
-inline u32 scaledown(u32 coef, u32 size)
+/// Draw3D pipeline step
+void Draw3D::run(PipelineContext *context)
 {
-	return (size + coef - 1) / coef;
+	if (m_target)
+		m_target->activate(context);
+
+	context->device->getSceneManager()->drawAll();
+	context->device->getVideoDriver()->setTransform(video::ETS_WORLD, core::IdentityMatrix);
+	if (!context->show_hud)
+		return;
+	context->hud->drawBlockBounds();
+	context->hud->drawSelectionMesh();
+	if (context->draw_wield_tool)
+		context->client->getCamera()->drawWieldedTool();
+}
+
+void DrawHUD::run(PipelineContext *context)
+{
+	if (context->show_hud) {
+		if (context->shadow_renderer)
+			context->shadow_renderer->drawDebug();
+
+		if (context->draw_crosshair)
+			context->hud->drawCrosshair();
+
+		context->hud->drawHotbar(context->client->getEnv().getLocalPlayer()->getWieldIndex());
+		context->hud->drawLuaElements(context->client->getCamera()->getOffset());
+		context->client->getCamera()->drawNametags();
+		auto mapper = context->client->getMinimap();
+		if (mapper && context->show_minimap)
+			mapper->drawMinimap();
+	}
+	context->device->getGUIEnvironment()->drawAll();
+}
+
+
+void MapPostFxStep::setRenderTarget(RenderTarget * _target)
+{
+	target = _target;
+}
+
+void MapPostFxStep::run(PipelineContext *context)
+{
+	if (target)
+		target->activate(context);
+
+	context->client->getEnv().getClientMap().renderPostFx(context->client->getCamera()->getCameraMode());
+}
+
+void RenderShadowMapStep::run(PipelineContext *context)
+{
+	// This is necessary to render shadows for animations correctly
+	context->device->getSceneManager()->getRootSceneNode()->OnAnimate(context->device->getTimer()->getTime());
+	context->shadow_renderer->update();
+}
+
+// class UpscaleStep
+
+void UpscaleStep::run(PipelineContext *context)
+{
+	video::ITexture *lowres = m_source->getTexture(0);
+	m_target->activate(context);
+	context->device->getVideoDriver()->draw2DImage(lowres,
+			core::rect<s32>(0, 0, context->target_size.X, context->target_size.Y),
+			core::rect<s32>(0, 0, lowres->getSize().Width, lowres->getSize().Height));
 }
 
 RenderingCorePlain::RenderingCorePlain(IrrlichtDevice *_device, Client *_client, Hud *_hud) : 
@@ -51,15 +118,4 @@ void RenderingCorePlain::createPipeline()
 		pipeline->addStep(pipeline->own(upscale));
 	}
 	pipeline->addStep(pipeline->own(new DrawHUD()));
-}
-
-// class UpscaleStep
-
-void UpscaleStep::run(PipelineContext *context)
-{
-	video::ITexture *lowres = m_source->getTexture(0);
-	m_target->activate(context);
-	context->device->getVideoDriver()->draw2DImage(lowres,
-			core::rect<s32>(0, 0, context->target_size.X, context->target_size.Y),
-			core::rect<s32>(0, 0, lowres->getSize().Width, lowres->getSize().Height));
 }
