+++ src/client/render/plain.h
@@ -20,6 +20,51 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #pragma once
 #include "core.h"
+#include "pipeline.h"
+
+/**
+ * Implements a pipeline step that renders the 3D scene
+ */
+class Draw3D : public RenderStep
+{
+public:
+	virtual void setRenderSource(RenderSource *) override {}
+	virtual void setRenderTarget(RenderTarget *target) override { m_target = target; }
+
+	virtual void reset(PipelineContext *context) override {}
+	virtual void run(PipelineContext *context) override;
+
+private:
+	RenderTarget *m_target {nullptr};
+};
+
+/**
+ * Implements a pipeline step that renders the game HUD
+ */
+class DrawHUD : public RenderStep
+{
+public:
+	virtual void setRenderSource(RenderSource *) override {}
+	virtual void setRenderTarget(RenderTarget *) override {}
+
+	virtual void reset(PipelineContext *context) override {}
+	virtual void run(PipelineContext *context) override;
+};
+
+class MapPostFxStep : public TrivialRenderStep
+{
+public:
+	virtual void setRenderTarget(RenderTarget *) override;
+	virtual void run(PipelineContext *context) override;
+private:
+	RenderTarget *target;
+};
+
+class RenderShadowMapStep : public TrivialRenderStep
+{
+public:
+	virtual void run(PipelineContext *context) override;
+};
 
 /**
  * UpscaleStep step performs rescaling of the image
