+++ src/client/render/stereo.h
@@ -20,6 +20,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #pragma once
 #include "core.h"
+#include "plain.h"
 #include "pipeline.h"
